requirements.txt
yfinance
nsetools
pandas
matplotlib
requests
beautifulsoup4
newspaper3k
praw
nltk
spacy
textblob
vaderSentiment
scikit-learn
torch
transformers
streamlit
plotly
flask
dash 
joblib
shap


